+++ __docs__/wonder-blocks-button/button-testing-snapshots.stories.tsx
@@ -22,7 +22,7 @@ export default {
             modes: themeModes,
         },
     },
-    tags: ["!autodocs"],
+    tags: ["!autodocs", "!manifest"],
     args: {
         children: "Button",
         icon: paperPlaneIcon,