+++ __docs__/wonder-blocks-tabs/responsive-navigation-tabs-testing-playtesting.stories.tsx
@@ -33,7 +33,7 @@ export default {
     args: {
         onLayoutChange: fn(),
     },
-    tags: ["!autodocs"],
+    tags: ["!autodocs", "!manifest"],
     globals: {
         theme: "thunderblocks",
     },