+++ __docs__/wonder-blocks-typography/deprecated-footnote.stories.tsx
@@ -1,0 +1,36 @@
+import * as React from "react";
+import {Meta, StoryObj} from "@storybook/react-vite";
+import packageConfig from "../../packages/wonder-blocks-typography/package.json";
+
+import ComponentInfo from "../components/component-info";
+import TypographyArgTypes from "./typography.argtypes";
+
+import {Footnote} from "@khanacademy/wonder-blocks-typography";
+
+export default {
+    title: "Packages / Typography / Deprecated / Footnote",
+    component: Footnote,
+    parameters: {
+        componentSubtitle: (
+            <ComponentInfo
+                name={packageConfig.name}
+                version={packageConfig.version}
+            />
+        ),
+        chromatic: {
+            // Disabling because all typography components are covered together
+            // in the Typography stories
+            disableSnapshot: true,
+        },
+    },
+    argTypes: TypographyArgTypes,
+    tags: ["deprecated", "!manifest"],
+} as Meta<typeof Footnote>;
+
+type StoryComponentType = StoryObj<typeof Footnote>;
+
+export const Default: StoryComponentType = {
+    args: {
+        children: "Footnote",
+    },
+};