+++ packages/wonder-blocks-accordion/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @khanacademy/wonder-blocks-accordion
 
+## 3.1.53
+
+### Patch Changes
+
+- Updated dependencies [358eb9b]
+    - @khanacademy/wonder-blocks-tokens@16.1.0
+    - @khanacademy/wonder-blocks-clickable@8.1.5
+    - @khanacademy/wonder-blocks-icon@5.3.9
+    - @khanacademy/wonder-blocks-typography@4.2.31
+
+## 3.1.52
+
+### Patch Changes
+
+- Updated dependencies [c75abe9]
+- Updated dependencies [c75abe9]
+- Updated dependencies [530d692]
+    - @khanacademy/wonder-blocks-tokens@16.0.0
+    - @khanacademy/wonder-blocks-clickable@8.1.4
+    - @khanacademy/wonder-blocks-icon@5.3.8
+    - @khanacademy/wonder-blocks-typography@4.2.30
+
 ## 3.1.51
 
 ### Patch Changes