+++ packages/wonder-blocks-breadcrumbs/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @khanacademy/wonder-blocks-breadcrumbs
 
+## 3.2.13
+
+### Patch Changes
+
+- Updated dependencies [358eb9b]
+    - @khanacademy/wonder-blocks-tokens@16.1.0
+
+## 3.2.12
+
+### Patch Changes
+
+- Updated dependencies [c75abe9]
+- Updated dependencies [c75abe9]
+- Updated dependencies [530d692]
+    - @khanacademy/wonder-blocks-tokens@16.0.0
+
 ## 3.2.11
 
 ### Patch Changes