+++ packages/wonder-blocks-styles/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @khanacademy/wonder-blocks-styles
 
+## 0.2.40
+
+### Patch Changes
+
+- Updated dependencies [358eb9b]
+    - @khanacademy/wonder-blocks-tokens@16.1.0
+
+## 0.2.39
+
+### Patch Changes
+
+- Updated dependencies [c75abe9]
+- Updated dependencies [c75abe9]
+- Updated dependencies [530d692]
+    - @khanacademy/wonder-blocks-tokens@16.0.0
+
 ## 0.2.38
 
 ### Patch Changes