+++ packages/wonder-blocks-toolbar/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @khanacademy/wonder-blocks-toolbar
 
+## 5.1.48
+
+### Patch Changes
+
+- Updated dependencies [358eb9b]
+    - @khanacademy/wonder-blocks-tokens@16.1.0
+    - @khanacademy/wonder-blocks-typography@4.2.31
+
+## 5.1.47
+
+### Patch Changes
+
+- Updated dependencies [c75abe9]
+- Updated dependencies [c75abe9]
+- Updated dependencies [530d692]
+    - @khanacademy/wonder-blocks-tokens@16.0.0
+    - @khanacademy/wonder-blocks-typography@4.2.30
+
 ## 5.1.46
 
 ### Patch Changes