+++ packages/wonder-blocks-typography/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @khanacademy/wonder-blocks-typography
 
+## 4.2.31
+
+### Patch Changes
+
+- Updated dependencies [358eb9b]
+    - @khanacademy/wonder-blocks-tokens@16.1.0
+
+## 4.2.30
+
+### Patch Changes
+
+- Updated dependencies [c75abe9]
+- Updated dependencies [c75abe9]
+- Updated dependencies [530d692]
+    - @khanacademy/wonder-blocks-tokens@16.0.0
+
 ## 4.2.29
 
 ### Patch Changes